Add files via upload
In this repository, you will find a jar file with the name Lab 10 that has the complete source files and code files. It contains a program which is capable of finding the size of the word,  count the letters of the longest word, count the words. Feel free to use.